pardus-ai-agent/requirements.txt
google-generativeai
colorama
gitpython